Remove space from PATH setting in /etc/environment
Issue #4

ComputeNode.sh and Scheduler.sh create /etc/environment with
a space at the end of the PATH value. This results in a corrupt
PATH value on the compute node, where it contains a double-quote
and a space. This patch removes the trailing space from PATH's
value in /etc/environment.